Refactor: LCFS - Decompose OrganizationsService into focused service classes (#4084)
Split the 1,443-line OrganizationsService class into focused domain services:
- OrganizationPenaltyService for penalty analytics and log management
- OrganizationLinkKeyService for form link key generation/validation
- OrganizationCreditMarketService for credit market updates, listings,
  audit logs, and notifications

OrganizationsService now focuses on org CRUD, balances, and the
adjust_balance orchestration that other modules depend on. Views inject
the focused services directly for the relevant routes; API contracts
are unchanged.
